minor cleanup based on feedback in review request
